Add support for checking whether an element is enabled and for verifying its attributes.
Add support for getting the current value of the url or of whether an
element exists.  This feature should be used sparingly, but sometimes it
is needed to get the app into a known state when a test is run in an
environment with shared resources.

Add a preflight and a postflight step to flight plans for tower setup
and teardown that is not associated with a test action.  Now that we
have preflight, add back the initial verfication before flying the first
leg of the flight plan.